+++ src/app/item-review/reducer.ts
@@ -3,6 +3,8 @@ import * as actions from './actions';
 import { ActionType, getType } from 'typesafe-actions';
 import { D2RatingData } from './d2-dtr-api-types';
 import { D1RatingData } from './d1-dtr-api-types';
+import { DimItem } from '../inventory/item-types';
+import { getReviewKey, getD2Roll } from '../destinyTrackerApi/d2-itemTransformer';
 
 // TODO: Should this be by account? Accounts need IDs
 export interface ReviewsState {
@@ -40,4 +42,22 @@ function ratingsFromItemStores(
     ratings[`${itemStore.referenceId}-${itemStore.roll}`] = itemStore;
   }
   return ratings;
+}
+
+export function getRating(
+  item: DimItem,
+  ratings: ReviewsState['ratings']
+): D2RatingData | D1RatingData | undefined {
+  let roll: string | null = null;
+
+  if (item.isDestiny1() && item.talentGrid) {
+    roll = item.talentGrid.dtrRoll;
+  } else if (item.isDestiny2()) {
+    const reviewKey = getReviewKey(item);
+
+    roll = getD2Roll(reviewKey.availablePerks);
+  }
+
+  const itemKey = `${item.hash}-${roll}`;
+  return ratings[itemKey] && ratings[itemKey];
 }