+++ docs/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next
 
 * We no longer support searching D1 vendor items.
+* Added support for showing ratings and reviews based on the item roll in Destiny 2.
 
 # 4.74.0 (2018-10-14)
 
@@ -8,7 +9,7 @@
 * Added `perk:"* **"` seach filter to match any keywords against perks on an item
 * Added some missing `stat:`
 * Lock and unlock items matching your current search from the same menu you use for tagging them.
-* Updated icons across the app.d
+* Updated icons across the app.
 
 # 4.73.0 (2018-10-07)